refactor(runtime): split tool_primitives + safety_net into focused packages
`runtime/tool_primitives.py` (1931 lines) and `runtime/safety_net.py`
(889 lines) had grown into kitchen-sink modules. The orchestrator-tools
review (top-10 findings) called both out as god modules; this PR
addresses the two highest-impact items.

== src/aise/tools/ (replaces runtime/tool_primitives.py) ==

11 focused modules with one job each:

- `context.py` — ToolContext + WorkflowState dataclasses
- `_common.py` — _now / _default_processes_dir
- `retry.py` — context-augmented retry prompt + constants
- `artifacts.py` — _artifact_shortfalls + size threshold
- `stack_contract.py` — load docs/stack_contract.json + UI/framework recipes
- `task_descriptions.py` — three deterministic worker-prompt renderers
- `discovery.py` — make_discovery_tools (list_processes / get_process / list_agents)
- `dispatch.py` — make_dispatch_tools (dispatch_task / .._parallel / .._subsystems)
- `shell.py` — make_shell_tool (allowlist-guarded execute_shell)
- `completion.py` — make_completion_tool (gated mark_complete)
- `builder.py` — build_orchestrator_tools aggregator

`runtime/project_session.py` switches to `from ..tools import …`. The
public surface (`ToolContext`, `WorkflowState`, `build_orchestrator_tools`)
is unchanged. Tests under `tests/test_runtime/` repoint their imports
to the new modules and the file-content assertion in
`test_subsystem_layout.py` is updated to read the new files.

== src/aise/safety_net/ (replaces runtime/safety_net.py) ==

Switches from a flat module of free functions to a registry + gateway
pattern, with checks and repairs grouped **by domain** so the strongly
coupled detection / remediation pair lives in one file.

Infrastructure:
- `types.py` — ExpectedArtifact, CheckOutcome
- `registry.py` — four registries (repairs / Layer-A invariants /
  artifact-kind handlers / artifact-repair policies) + decorators
- `gateway.py` — single public entry point `run_post_step_check`,
  routes via the registries; no more `if name == "..."` chains
- `events.py` — telemetry writer
- `expectations.py` — three pre-baked ExpectedArtifact sets
- `repair_policy.py` — ExpectedArtifact → repair name mapping

Domains (each module owns BOTH its checks and its repairs):
- `git.py` — VCS surface: git + .gitignore. Holds the git-repo and
  gitignore invariants, the git_repo / git_tag / clean_tree artifact
  handlers, the git_init / autocommit / phase_tag / seed_gitignore
  repairs, the shared `_run_git` helper, and the baseline `.gitignore`
  content. One file = every git-touching action.
- `filesystem.py` — standard-subdirs invariant + the file / dir /
  json_file / must_not_exist artifact handlers + the
  create_standard_subdirs and remove_leftover repairs.
- `stack_contract.py` — JSON-schema validator + handler. No repair:
  malformed contract is healed by re-dispatching the architect.

Each domain module uses decorators (`@register_repair("…")`,
`@register_invariant("…", category="scaffold")`,
`@register_artifact_kind("…")`, `@register_artifact_repair_policy`)
so the gateway never needs to know about specific symbols. Adding a
new check + repair pair is one new module + one branch in
`repair_policy.py`.

Public surface preserved: `run_post_step_check`, `ExpectedArtifact`,
`CheckOutcome`, `scaffolding_expectations` / `architecture_expectations`
/ `qa_expectations`, `REPAIR_ACTIONS`, `LAYER_A_INVARIANTS`,
`_stack_contract_valid`, `_artifact_present`, `_BASELINE_GITIGNORE`
all re-exported from `aise.safety_net`. `REPAIR_ACTIONS` shares the
registry's underlying dict directly so `monkeypatch.setitem(...)`
in tests still reaches the gateway lookup.

== Migration ==

- `runtime/project_session.py`: import path change
- `runtime/policy_backend.py`: docstring reference
- `web/app.py`, `web/safety_net_events_service.py`, `web/static/app.js`:
  import path + docstring references
- 5 runtime tests + 2 web tests: import path updates only (no
  behaviour change)
- `runtime/tool_primitives.py` and `runtime/safety_net.py` deleted

Validation: ruff check / ruff format --check / scripts/validate_skills_registry.py /
pytest (576 passed, 50 skipped) / `python -m build` all pass.
